wms: refactor time management
- remove duplicate
- more code from worker to utils
- add test
