+++ package.json
@@ -126,6 +126,7 @@
   "devDependencies": {
     "eslint": "8.37.0",
     "eslint-config-next": "13.2.4",
+    "jest": "^30.3.0",
     "sass": "^1.60.0"
   },
   "cacheDirectories": [
